% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfn.R
\name{clean_logger_name}
\alias{clean_logger_name}
\title{Build a 'clean' logger name}
\usage{
clean_logger_name(x)
}
\arguments{
\item{x}{\code{chr(n)}}
}
\value{
\verb{chr(n* <= n)}
}
\description{
Build a 'clean' logger name
}
\details{
Given a vector \code{x} of length \code{n}, this function:
\enumerate{
\item coerces to \code{chr(n)} (and throws an error if non-coercible),
\item whitespace-trims the new vector's elements, and
\item removes from the vector any empty or \code{NA} elements.
}
\subsection{Motivation}{

\code{lgr::get_logger(x)} will happily create loggers with names like \verb{/foo/  /bar/NA//baz} if the input is \code{c("foo", "  ", "bar", NA_character_, "", "baz")}.
This function applies some sanity to that process and converts that example input into \code{c("foo", "bar", "baz")}, which can then be forwarded-on to \code{lgr::get_logger()}.
}
}
