% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a logger with nice defaults}
\usage{
init(
  ...,
  lgr = lgr::lgr,
  threshold = "all",
  layout_fmt = LAYOUT_FMT,
  timestamp_fmt = TIMESTAMP_FMT
)
}
\arguments{
\item{lgr}{the \code{lgr::Logger} to initialize (i.e. modify in-place); {lgr}'s root logger i.e. \code{lgr::lgr} by default.}

\item{threshold}{the threshold to use for both the Logger \emph{and} all attached appenders; typically one of \code{all}, \code{trace}, \code{debug}, \code{warn}, \code{info}, or \code{fatal}.}

\item{layout_fmt}{the layout format string; by default \verb{\%L [\%t] [\%g] \%m \%f}.}

\item{timestamp_fmt}{the timestamp format string; by default \verb{\%FT\%T\%z} (ISO8601 datetime).}
}
\description{
Initialize a logger with nice defaults
}
\details{
All arguments must be named.

This function sets the threshold and this package's layout/timestamp formats on the specified Logger, which be default is {lgr}'s root Logger (i.e. \code{lgr::lgr}).

Like other functions in this package, this function cannot fail.
If invalid arguments are given, this function will warn but not err.
}
