% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfn.R
\name{logfn}
\alias{logfn}
\title{Create a new, scoped, logging function.}
\usage{
logfn(x = NULL, ..., context = list())
}
\arguments{
\item{x}{\code{chr(n >= 0) | NULL | Logger | logfn}}

\item{context}{a (typically-named) list providing key/val context spliced into \code{...} in the underlying logging call.}
}
\value{
a (possibly-new) scoped logging function.
}
\description{
Create a new, scoped, logging function.
}
\details{
If \code{x} is \code{NULL}, character vector, or a Logger object, we assume we're creating and/or wrapping a Logger.
If \code{x} is a single \code{logfn}, we're wrapping (or 'layering deeper within') an existing \code{logfn} wrapper.
In both cases, we may be adding context.

This logging function will always take the sgnature: \code{f(level, msg = "", ...)}, meaning:
\itemize{
\item \code{level} \strong{must} be specified (and exist for that Logger --- in the rare case where new log levels are created).
\item \code{...} hold the spliced-in \code{context}, to be used when interpolating into \code{msg}.
}

The logging function will \strong{never fail}.
If it's called errantly, it will warn and swallow any error encountered during the call.
This helps enforce the principle that adding logging should never introduce errors into the surrounding program.
}
