% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfn.R
\name{logfn_logger}
\alias{logfn_logger}
\title{Get a logfn's underlying Logger}
\usage{
logfn_logger(logfn)
}
\arguments{
\item{logfn}{the logging function.}
}
\value{
\code{logfn}'s underlying (delegated-to) Logger.
}
\description{
Get a logfn's underlying Logger
}
