% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfn.R
\name{logfn_msg}
\alias{logfn_msg}
\title{Converts input to a log-friendly message string}
\usage{
logfn_msg(msg, fail_msg = "<bad log msg>")
}
\arguments{
\item{`msg`}{the source message (or wrapper).}

\item{`fail_msg`}{the log message to use if \code{msg} can't be coerced into a log-friendly message (string).}
}
\value{
a log-friendly message (string).
}
\description{
Converts input to a log-friendly message string
}
\details{
If \code{msg} is a non-NA string (\code{chr(1)}), it is returned as-is. Notably, the empty string (\code{""}) is happily accepted here.
If \code{msg} is a condition, the condition message is extracted and returned.
Otherwise \code{fail_msg} is returned.
}
