% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murlib.lgr-package.R
\docType{package}
\name{murlib.lgr-package}
\alias{murlib.lgr}
\alias{murlib.lgr-package}
\title{murlib.lgr: Tools for {lgr}}
\description{
Some helpers for using the (excellent, but sometimes too-flexible) {lgr} package.
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/0000-0002-8155-2565}{ORCID})

}
\keyword{internal}
