% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfn.R
\name{string_repr.R6}
\alias{string_repr.R6}
\title{Generic (fallback) log message pretty-printer for R6 objects.}
\usage{
\method{string_repr}{R6}(x, width = 32L, ...)
}
\description{
Generic (fallback) log message pretty-printer for R6 objects.
}
\details{
To override, add \code{string_repr(x, width, ...)} method to an R6's \code{public} method list.
}
